+++ models/User.js
@@ -11,23 +11,7 @@ const userSchema = new mongoose.Schema(
       default: "Client",
       enum: ["Admin", "Driver", "Client", "Vendor"],
     },
-    dietType: {
-      type: String,
-      default: "High-Carb",
-      enum: ["High-Carb", "Low-Carb", "Vegan", "Ceto"],
-      required: true,
-    },
-    genders: {
-      type: String,
-      default: "male",
-      required: true,
-      enum: ["female", "male"],
-    },
-    dateOfBirth: { type: Date },
-    weight: { type: Number },
-    height: { type: Number },
-    healthInfo: { type: mongoose.Schema.Types.ObjectId, ref: "Healthinfo" },
-
+    healthInfo: { type: mongoose.Schema.Types.ObjectId, ref: "HealthInfo" },
     address: [{ type: mongoose.Schema.Types.ObjectId, ref: "Address" }],
     wishlist: [{ type: mongoose.Schema.Types.ObjectId, ref: "Recipe" }],
     phone: { type: String, required: false },
